Interpretable encoding scheme from dataset dimensions to data arrays. (#88)
Here, I've introduced dataclass that encodes and decodes a (mostly)
interpretable `xarray.DataArray` of floats. These are manually tested
with a fuzzer (hypothesis). These tests are turned off by default due to
the choice to encode the numbers as significant digits in floats due to
the tendency for rounding errors in `np.float64` values. For our
purposes of making debuggable test data, these should be decent.

Fixes #79.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>