Resolve CFTime warning (#139)
The underlying [warning
](https://github.com/Unidata/cftime/blob/eab0212df73decd5a420adaae7938b942d131c95/src/cftime/_cftime.pyx#L1151)
was due to us passing a date with year < 0 for a calendar that doesn't
support that.

This was for 2 reasons: first, we had no minimum in the possible
generated dates and second because the fromordinal function we were
using takes the number of days since year -4713 (yes, negative four
thousand years BCE) but the python datetime toordinal produces the
number of days since year 1. So, we have a minimum now and we are using
a conversion function which is a little less silly.

Fixes #112