March refactor - renaming (#141)
* Bunch of renaming and removal of unused experiment configs in
constants. The most important rename here is inputs -> prognostic, extra
-> boundary and removal of output variables names/keys since they are
the same as input variables. This functionality is okay for now since
the codebase depends on this fact. I don't see a near future where this
is going to change.
* The scripts for training on gantry, empire ai etc. may not work and I
am planning to fix that in a later PR

Following Draft PR #134 

Closes issue #80