Torch-first data loader without dask (#178)
I did some profiling and continue to see a huge amount of time spent in
xarray indexing. I built a dataloader which goes as quickly as possible
into (CPU) torch and does not use dask. On my laptop this is 3x faster
than the eager data loader, and on my test GPU machine it's about 5x
faster in the data loader benchmarks. Also seems to eliminate any
waiting for data on my GPU machine with 4 workers when running with
batch size == 2.