Introducing `DataSource` to train and eval scripts. (#211)
This makes progress towards #190. Specifically, this PR argues that the
three Xarray datasets (data, means, stds) are actually part of the same
concern, one that is fundamentally related to normalization. I
anticipate that this affordance will make it easier to modify core
aspects to the data loader in the future. In particular, the addition of
the `DataSource.filter()` method should provide a single entry point to
add data compaction support -- see #215 (initially prototyped in #175).
In addition, this change should make #208 trivial to implement.

Fixes #218.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>