Use dask for inference (#227)
Fixes memory issues with the torch loader, fixes #208. 

After this, the torch and eager loaders have the same peak memory usage
during inference. The issue was that the non-dask version was using
about 3x the memory of the dask version and getting OOM killed. This is
due to a combination of the InferenceDataset not being written with
eager evaluation in mind (eg slicing repeatedly) and due to what look
like some pretty bad performance bugs in xarray. (It was allocating a
ton of memory to lazily track reorderings of the data rather than
eagerly shuffling the arrays -- kinda looks like 6x more than the
underlying arrays but I could have misread and don't think it's worth
digging into this more now.) (And to add insult to injury, the shuffle
in question was the identity transform…)