Switch to using EC2 for CI workers, take 2 (#250)
This changeset switches the repo to use OpenAthena's EC2 account to
perform CI checks.

Unlike its predecessor, #246, this PR depends on a branch of
`start-gha-ec2-runner` which contains a fix for a bug that prevented the
runner machines from being detected by GHA. We will upstream this fix
and send a follow-up PR when the fix has been accepted.

We observe a modest improvement in benchmarks when running on AWS
`g6.xlarge` VMs vs `g2-standard-8` Google Cloud VMs.