Switch to torch loader and reduce default number of workers per GPU (#255)
I think the torch loader is a good default now, and I found 4 workers by
default overloads the 8-GPU systems I've been using, resulting in very
long waits at the start of an epoch. 1-worker-per-GPU seems to be too
few on my other host, so I was thinking this might be a better default.
(Though of course might need tuning.)