Use passed loader version to decide whether to use dask (#260)
@alxmrs pretty sure this is why our benchmarks did something surprising
when we changed the default loader version. Previously the passed
`version` was ignored for the purpose of deciding whether to use dask or
not. Seems like good additional evidence that we should try and tackle
this duplication sometime soon.

This also makes me want to move to a world where we're not using config
objects in most tests. It's hard to tell whether this change is
sufficient -- if any other code we're calling here transitively uses
`cfg.data.loader_version`, then it's probably broken. We could also make
a new train config but then if the *caller* is using
`cfg.data.loader_version` we have the same problem. We could update all
callers to mutate train config rather than passing in verison and
time_config separately but I would rather move away from manipulating
config rather than towards. (But could be convinced this is still worth
doing in the short term.)