Notes on how to reproduce OM4 half-deg preprocessing. (#261)
I've refreshed the contributing guide, adding a data engineering
section. Here, I explain how to clone the data and how to start a coiled
notebook to create the half-deg data. The docs indicate that this all
may change soon -- and that is my plan. I am happy to refresh these docs
overall.

---------

Co-authored-by: Surya <surya.dheeshjith@gmail.com>