Tweaks while running half-degree experiments (#283)
The 3 commits are independent so probably easiest to review that way.
(Happy to break into separate PRs but I think they're all pretty
uncontroversial.)

* Updates config for the half-degree experiments to load remotely when
desired and to avoid doing inference during training.
* Adds a flag to load xarray variables concurrently based on [this
suggestion](https://github.com/pydata/xarray/issues/8965#issuecomment-2083787484).
This helps a lot locally with small datasizes (3x improvement for me)
but with larger batch sizes it seems to not help (maybe GIL
contention?).
* Swaps to the the "spawn" multiprocessing type since the s3fs requires
this when passing those stores around.
* Adds more logging statements so you can tell what's going on more
easily
* Removes some logic we had about tweaking the number of workers based
on the number of GPUs in use -- the worker counts are are *already* per
GPU so this was using `num_gpus * num_gpus * num_workers` workers in
total. Also allows us to use multiple workers for CPU jobs, for testing
purposes.