Fix performance regression by moving towards a more buildy world with DataContainer (#295)
Fixes the performance regression from
[here](https://github.com/LaureZanna/Ocean_Emulator/commit/32f3043d20bb43b24a1a097004f5c0f34f013075#commitcomment-160869800)
by detecting if the data source supports forking or not.

Doing this required a bit of code movement so I took the opportunity to
move us closer towards a world where creating the trainer is `trainer =
cfg.build()` by moving a bunch of the stuff in `Trainer.__init__` into
`DataConfig.build` and creating a new class which is the result of that
method, `DataContainer`. This also reduces some of the duplication
between train/eval/tests.

Last commit is some drive-by fixes for warnings that were being printed
in tests.