Avoid LocalLocation relative footgun (#309)
Previously, we permitted you to make *relative* local locations as in:

```yaml
data_location:
  type: local
  path: foo/bar
```

But the rule is that if you give a ResolvedLocation (type: local is a
ResolvedLocation) then it's *not* resolved relative to the data_root. So
this is actually relative to the cwd rather than data_root. This is just
confusing so we are now forbidding relative local locations; instead you
want:

```yaml
data_location: "foo/bar"
```

This means we can no longer express "resolved paths which are relative
to cwd" but this seems preferable to having this footgun.

---------

Co-authored-by: Alex Merose <alex@openathena.ai>