Hopefully fix data race in cache reading/writing (#318)
Breaks out the test change from #313 and fixes the bug (see eg [this
actions
run](https://github.com/Open-Athena/Ocean_Emulator/actions/runs/16573115391/job/46880287790))
that exposed.

We now ensure that only one process (of possibly-many pytest processes)
write to a given data source cache at once. We also ensure that
processes that would like to `open` a given cache wait for any writers
of that cache to finish. (They also wait for other readers, though this
is unnecessary.) This locking is also hoisted upwards such that only one
process will try to create the remote/compact/mock data in the case when
it is not cached, which should hopefully reduce the chance of that
failing.

Fixes #257