Fix EMA state restore when resuming training from a checkpoint (#313)
This fixes a silly bug which overwrote all EMA state after resuming from
a checkpoint. It also now declares the two kinds of checkpoints to be
for different uses:

* EMA checkpoints: these bake in the averaged weights into the model
itself, and so are useful for inference that wants to start from that
averaged model.
* Non-EMA checkpoints: these now save the averaged model weights
separately from the main model weights so we can resume training without
starting that averaging over again.

These maybe could be renamed "inference checkpoints" and "training
checkpoints" or something if this turns out to be a useful distinction.

I ended up not adding a test that training continues exactly as it would
have without checkpointing/restoring because I think there are other
bugs that are not critical to solve right now like rand seeds not being
saved. (I started a [discussion
here](https://openathena.slack.com/archives/C0884476QSC/p1753229734048809))

Fixes #282