Initializing EMA after wrapping model with DDP
I encountered a bug where I hit a key error during the EMATracker call phase where it was unable to find the key for any layer.
```
 File "/home/ubuntu/sky_workdir/src/ocean_emulators/utils/ema.py", line 129, in __call__
    ema_name = self._module_name_to_ema_name[key]
                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^
  KeyError: 'module.layers.0.skip_module.weight'
```

I found that this was caused by wrapping the `DDP` component onto the model -- doing so adds a `module` prefix to the weights. Thus, the names of the weights don't match the model setup and EMATracker call during training.

In this PR, I make a simple fix, which is to just move the DDP wrapper code above the part where we set up the EMATracker.