Fix UnresolvedLocation JSON schema to permit string (#325)
This was previously causing red squigglies in VS Code. It's annoying
that this annotation and the two `@model_*` annotations on this class
are all required -- I spent some time looking for a more direct way to
express "this type serializes to/from a string" in the pydantic docs but
everything I tried was broken or more complicated than this.