Sky Train: Cloning data faster via setup section instead of file mount (#319)
Since `file_mount` options don't let you pass arguments to rclone, this
PR opts to manually `rclone` data during the `setup` stage of the job.
This allows us to make faster file transfers with the `--transfers`
flag.

It also has required that I map the `data/` dir to home and not root. 
We also set the default number of workers to 4.