Implementing Warmup + Cosine Decay LR schedule from ConvNeXt paper. (#343)
Here is a Warmup implementation. This was added to match the LR schedule
of the original ConvNeXt paper. Following the citations, warmup is
theorhetical grounded and empirically verified to help the loss curves
of networks with _large batch sizes_ -- see this paper:
https://arxiv.org/pdf/1706.02677

We have very small batch sizes, though each sample is fairly large (~80
MB). This technique is not likely to lead to a modeling improvement;
however, I've queued up a few jobs to test if this improves anything.
While this might not make it into the final baseline Samudra config, I
could see it being valuable for the next iteration of the model.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>