Updated our configuration system to accommodate multiple models, including FOMO v0.  (#379)
This change makes space for us to develop and configure non-Samudra
models. I've created a basic implementation of FOMOv0, but this is
intentionally a placeholder. The focus of the PR is on the config system
changes. To that end, all models are constructed with our builder
pattern.

For the benefit of working with both FOMOv0 and Samudra, this PR factors
out the UNet implementation into its own module. We've also updated all
our config files to follow this new interface.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>