FOMO module components agree in dimension.
I'm following an `in_channel`/`out_channel` pattern for all Modules now. I've modified the config system to `build()` with these two values as inputs. The FOMO model now accepts two tunable parameters: `in_embedding` and `out_embedding`, which define the size of the latent channels of the processor.
The UNet now accepts `in_channels` and `out_channels`. New in this version: we can set a value for `out_channels` that differ from the `ch_widths` -- we are backwards compatible with Samudra.

The trainer test now includes testing a FOMO config.