Making the encoder more compatible with the processor.
The UNet expects the channel dimension to be on axis=1, but our encoder assumes it's the last channel. In addition, I've fixed a minor bug where we want to apply layer norm before reshaping instead of after.

Last, I've set a few hparams for the Perceiver that I'm definitely sure we'll want, namely weight sharing!