First round of perf improvements  (#392)
3 small fixes from profiling:

* Request async data transfers from CPU to GPU memory.
* Add the
[required](https://docs.pytorch.org/docs/stable/data.html#memory-pinning)
`pin_memory` method on our custom data type so those transfers can be a
DMA
* Remove a number of places we waited unnecessarily for GPU results to
be ready (to check against NaN) which caused our CPU to wait there for
the GPU to catch up. This isn't very helpful yet but is a step towards
freeing up the CPU for other things (eg data decompression).