Allow us to block ec2-related failures from merging (#423)
#420 was merged without GPU tests passing. This was permitted by our
branch protection rules because the "run-tests" job in the GPU tests
workflow was *skipped* (because one of its dependencies, the ec2 job
which spins up a GPU for us) failed. Perhaps surprisingly, this skipped
state [counts as
success](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks)
for the purpose of branch protection rules.

So this defines a new job which explicitly fails when either the
ec2-spin-up or GPU test themselves fail, which we'll add to our branch
protection rules.