Multiple, configurable perceiver implementations, including with flash attention. (#426)
This PR extracts the Perceiver into its own config. This allows us to
specify which perceiver implementation to adopt in either the encoder or
decoder (just the encoder for now). We add a flash attention perceiver
as the second option beyond a "naive" perceiver. Last, users can select
"auto" as an implementation option, and the config will pick the best
implementation given the runtime environment.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>