Move data loader compute pieces to GPU (#425)
We were spending a ton of CPU time doing normalization, which was
causing all kinds of bad knock-on performance effects due to (a) not
having CPU time for the main process and (b) data loaders not keeping up
with the GPU since they were busy doing math.

This wraps the torch DataLoader with a new class which knows how to do
some post-processing on the GPU before handing off the data. This keeps
the I/O and xarray/zarr/etc python overhead in separate processes but
lets us do the normalization on the main process on the GPU. Experiments
on half-degree data shows this cuts train iteration time significantly.

To avoid having multiple paths here, this also deletes the old "eager"
data loader. I'd also like to delete the inference data loader to reduce
duplication caused by it but figured it could be a follow-up.

Closes #412 and #403

---------

Co-authored-by: Alex Merose <alex@openathena.ai>