Remove TIME_DELTA constant and compute inference steps from data directly (#471)
The underlying bug in #468 was that the TIME_DELTA was set to 5 but the
actual time between individual steps in @mkeutgen's dataset was 1 day.
This lead to a mismatch between the size of the array the inference
aggregators were accumulating into and the actual produced data's shape.

This is the only place that constant was still used and it seems like we
can compute the steps more directly from the data so this kind of error
can't happen again.

@mkeutgen if you wouldn't mind trying out this change with your data
that would be a great sanity check!

Closes #468