Refactoring dynamic loss to compose any loss method, not just MSE. (#464)
We now use a builder function to create the loss function. Loss can be
defined as a simple metric, or a more sophisticated loss mixed in with a
metric, such as dynamic loss. This should allow us to combine MAE with
dynamic loss, for example.

My hope with this change is to follow up with experimenting with this
idea from the Cambdige/NVIDIA paper:
https://ar5iv.labs.arxiv.org/html/2410.07472v1#S4.SS10.

(TL;DR: I have reason to believe that dynamic MAE will perform better
than dynamic MSE).