Various minor viz doc & config fixes (#502)
This is a bit awkward because the viz configs don't really have anything
to do with the folder they're in (other than trying to be helpful by
having the source eval run name included, though arguably this is
confusing more than helpful). And really they vary mostly by what
resolution of data you need (and so data path + basin mask path). But I
think this is at least improving the situation a bit by fixing typos and
defaulting everything to 1-degree data.