Masking is same concern as the DataSource. (#499)
Introducing a `Masks` dataclass the in `DataSource` which has extracted
wet masks. It turns out, near everywhere that `DataSource` is used, we
have to separately plumb the surface and default wetmask. Thinking about
what it would take to load multiple datasets of different resolutions,
it seems that ocean/land masks are actually part of the same concern in
as the data source. My hope is that after this PR, we can introduce a
separate, but related masking concept -- namely, a dataloader-time
"nodata" mask that is only necessary when combining disparate datasets.

Along the way, this refactor also brings in data validation into the
creation of the `DataSource`.

---------

Co-authored-by: Jesse Rusak <jesse@openathena.ai>